cancel state divid, moidy delete handle
